Update app_config.php
Quote from Jonathan Black
https://www.pbxforums.com/threads/working-sms-integration.2675/page-3
It looks like something is wrong in the \sms\app_config.php file. It's looking for two tables named "v", not finding them, and trying to create them. Instead, it should be looking for "v_sms_messages" and "v_sms_destinations". There may be updates to the 4.4 schema upgrade engine. I've tweaked a couple of settings in the attached. 

Tested the above and it resolves this issue.